+++ src/it/test-projects/creedengo-java-plugin-test-project/src/main/java/org/greencodeinitiative/creedengo/java/checks/UseFilterBeforeSort.java
@@ -1,0 +1,27 @@
+package org.greencodeinitiative.creedengo.java.checks;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+class UseFilterBeforeSort {
+    UseFilterBeforeSort() {
+    }
+
+    public void manipulateStream(final List<String> list) {
+        list.stream() // Noncompliant {{Use 'filter' before 'sorted' for better efficiency.}}
+                .sorted()
+                .filter(s -> s.startsWith("A"))
+                .collect(Collectors.toList());
+
+        list.stream() // Noncompliant {{Use 'filter' before 'sorted' for better efficiency.}}
+                .sorted()
+                .map(element -> element.toString())
+                .filter(s -> s.startsWith("A"))
+                .collect(Collectors.toList());
+
+        list.stream() // Compliant {{Use 'filter' before 'sorted' for better efficiency.}}
+                .filter(s -> s.startsWith("A"))
+                .sorted()
+                .collect(Collectors.toList());
+    }
+}